# JSON.parse optimizations, refactoring, DoS-attack prevention