--- a/Jint/Native/Array/ArrayConstructor.cs
+++ b/Jint/Native/Array/ArrayConstructor.cs
@@ -366,15 +366,10 @@ namespace Jint.Native.Array
 
         public JsArray ConstructFast(JsValue[] contents)
         {
-            return ConstructFast(contents, 0, contents.Length);
-        }
-
-        internal JsArray ConstructFast(JsValue[] contents, int offset, int length)
-        {
-            var instance = ArrayCreate((ulong) length);
-            for (var i = 0; i < length; i++)
+            var instance = ArrayCreate((ulong) contents.Length);
+            for (var i = 0; i < contents.Length; i++)
             {
-                instance.SetIndexValue((uint) i, contents[i + offset], updateLength: false);
+                instance.SetIndexValue((uint) i, contents[i], updateLength: false);
             }
             return instance;
         }