--- a/Jint/Options.Extensions.cs
+++ b/Jint/Options.Extensions.cs
@@ -213,12 +213,11 @@ namespace Jint
             return options;
         }
 
-        public static Options MaxJsonParseDepth(this Options options, uint maxDepth)
+        public static Options MaxJsonParseDepth(this Options options, int maxDepth)
         {
-            options.Constraints.MaxJsonParseDepth = maxDepth;
+            options.Json.MaxParseDepth = maxDepth;
             return options;
         }
-
 
         public static Options SetReferencesResolver(this Options options, IReferenceResolver resolver)
         {