@tomatosalat0 2023-03-12
Hi,

thanks for taking your time to review it. Will go through the comments and adjust the PR accordingly. Might take a day or two, depending on when I find some time to do it.